SMALL EDITS AND CONDENSE SEMANTICS SECTION AT END
Small edits and changes for style etc.

This is for the oxford application, and the essays need to be 2000-2500
words, and mostly critical (not expository). I think the part at the end
about the limitations this places on semantic theories is good, but
could do with being shorter and clearer, so begin cutting it down.

Comment out next section to cut down